requirements.txt
attrs==21.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
docx==0.2.4
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pydocstyle==6.1.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
